+++ app/(public)/[role]/auth/find/hooks/useFindId/index.ts
