+++ app/(public)/[role]/auth/find/hooks/useResetPassword/index.ts
